Fixes for win32 - minor

+++ src/simplecsv.cpp
@@ -95,7 +95,7 @@ void Row::set(size_t cidx,const std::string &value)
 
 void Row::dump() const // {{{
 {
-  printf("[%lu]:",ridx);
+  printf("[%llu]:",(unsigned long long)ridx);
 
 //  for (std::map<size_t,Value *>::iterator it=columns.begin(),end=columns.end();it!=end;++it) {
   const size_t clen=size();
